+++ src/core/server/systems/toolbar.ts
@@ -28,6 +28,12 @@ export class ToolbarController {
             return;
         }
 
+        // Handle Consume Item from Toolbar
+        if (isFlagEnabled(item.behavior, ItemType.CONSUMABLE)) {
+            ToolbarController.handleToolbarUse(player, item);
+            return;
+        }
+
         // Handle other item switch types
         // No idea what this will be yet.
     }
@@ -67,6 +73,24 @@ export class ToolbarController {
 
         player.lastToolbarData.equipped = false;
         playerFuncs.emit.sound3D(player, 'item_remove', player);
+    }
+
+    static handleToolbarUse(player: alt.Player, item: Item) {
+        item.quantity -= 1;
+
+        if (item.quantity <= 0) {
+            playerFuncs.inventory.toolbarRemove(player, item.slot);
+        } else {
+            playerFuncs.inventory.replaceToolbarItem(player, item);
+        }
+
+        if (item.data && item.data.event) {
+            alt.emit(item.data.event, player, item);
+        }
+
+        playerFuncs.save.field(player, 'inventory', player.data.inventory);
+        playerFuncs.sync.inventory(player);
+        playerFuncs.emit.sound2D(player, 'item_use', Math.random() * 0.45 + 0.1);
     }
 }